issue-29 Remove unused file
